Fix login/signup issue
Issue was with the Catalyze API key, which needs “ios catalyze” in front of what I had. Now users can sign up and login. Users are segued to the main screen when this occurs.
Confirmed with Catalyze CEO that there is no data dashboard like there is on Parse—we’d have to create our own with the API. Will investigate this to try to confirm active users and record a test entry. From there, could use ACLs to restrict data access.