Add support for pain entering
Create symptom bit mask array of ints. Populate it with arbitrary values in a loop. Create cycleSymptoms method which uses switch to determine which method to call. Call painMethod in case 0 which creates and operates UI elements for pain recording. No action taken on submit (dependent on backend). Should consider moving painMethod and other methods to other classes if possible