Add confirm alert
Restructured program so submit pressed just shows an alert. The main view controller is the delegate for the alert view. When no is pressed, it returns to the same view. The code to move on to the next symptom is now moved to the case where the user pressed “yes” to indicate input is correct.