Minor changes for querying and saving
Adjust saving in background to be creating in background so Catalyze saving actually works!!! Move NSMutableDictionary previousDictionary to be a class variable, as well as the appDel PCAAppDelegate reference. Move some stuff in viewDidLoad so that querying for previous dictionary values is done before UI elements are created. Adjust UI code to use dictionary value to display previous score, though due to querying issues this does not work. Will attempt to update Catalyze SDK to see if that helps with querying.