Fix query issue
Query issue was not really an issue at all. The query executes asynchronously, so I was hitting the following code (which attempted to use a nil Dictionary) before the query finished.

Solution:
Prepare the UI label initially with a question mark. Once the query is done, in the callback, call a new method to update that UI element.
The reference to that element is now global and should be fixed with the rest of the global elements probably

NOTE: This still uses the first result from the query as the most recent. This is not always the case!!!