Add query helper to determine most recent entry
Added findMostRecent to determine what the most recent entry was for showing the user’s last entered value was.
According to email from Josh Ault, database results are deterministic and are very close to sorted by createdDate (but not guaranteed-determined by database). As a result, we can be virtually assured that the result will come in the 100 page size specified (or even in 50, probably—possible future change).

TODO: doctor stuff