Change to support double saving rather than int
In response to feedback from Lelan, determined that values should be saved to the tenths place. Updated app to support doubles. Updated custom class on Catalyze. Saving round numbers does not currently work--NSNumber storing 5.0 stores it as 5, which throws an error when saving to a double on Catalyze. This should not occur--widening should be valid--so I contacted Catalyze to inquire about the error.