Adjust behavior when finished
Goal is to prompt patient thrice weekly for information. Added function which will eventually check whether patient has already entered three times this week or not. If the patient has, it sets a value on the AllDone view controller to display that information. Otherwise, the AllDone controller will continue to represent the state when users have finished entering symptoms.