Remove symptom bit mask and adjust all done behavior
App is in reasonable shape but users still see "Main page" come up before the "All done" screen if they've already entered symptoms.
Fixing this would require reordering the app flow so that the shouldCycleSymptoms check is done in viewwillappear or something, to skip the view. Still, might not solve it. Worth solving?