Kinda implemented a segue
Glitchy selection--only calls "did select" the second time. Look into this more. Also, sort the array by numUrgent first rather than calculating that on the fly as you render the cells, since you want urgent patients at the top. Once you figure that out, it should look pretty good. Note that viewdidload in the detail controller is not changing the title so it might not be passing the information correctly. One option is to send the cell instead of "self" in the perform segue.